Refactor preset data modules
