+++ docs/index.html
@@ -473,7 +473,6 @@
       </main>
     </div>
 
-    <script src="./js/layout-presets.js"></script>
     <script type="module" src="./js/app.js"></script>
   </body>
 </html>
