+++ docs/js/data/input-presets.js
